実効時間の計測処理を追加

+++ Leibniz.java
@@ -3,7 +3,7 @@ import java.lang.*;
 
 public class Leibniz {
 	
-	public double calc(int num){
+	public double calc(long num){
 
 		//‰~ü—¦
 		double pi_result = 0.0;
@@ -16,12 +16,13 @@ public class Leibniz {
 		}
 
 		//ŒvZƒƒCƒ“
-		for(int i = 0; i < num; i++){
+		for(long i = 0; i < num; i++){
 			//ŒvZ
 			pi_result += operator / (2.0 * (double)i + 1.0);
 			//•„†”½“]
 			operator *= (-1.0);
 		}
+
 		return (pi_result * 4.0);
 	}
 }
